Introduce NamedXContentRegistry
Introduces NamedXContentRegistry which Works like
`NamedWriteableRegistry`: on startup components register parse
methods under names and a superclass. At runtime we use the
`XContentParser.namedXContent` method to lookup the parser and
parse the object.

Right now the parsers take a `context` object they use to help
with the parsing but I hope to be able to eliminate the need
for this context as most what it is used for at this point is
to move around parser registries which should be replaced by
this method eventually.

This converts `ScoreFunction` parsing to this method because it
is relatively self contained.